acquisition operational, fft bad results

+++ samv71_softpack_1.5_for_ewarm_7.30/arm/examples/Atmel/SAMV71_Xplained_Ultra/examples/spi/pc_dma.h
@@ -18,7 +18,7 @@ typedef struct {
 	/** Pointer to the Rx data. */
 	uint32_t *pRxBuff;
 	/** Rx size in bytes. */
-	uint16_t RxSize;
+	uint32_t RxSize;
 	/** Callback function invoked at the end of transfer. */
 	PcCallback callback;
 	/** Callback arguments. */
